added tests

+++ app/routers/users.py
@@ -2,6 +2,7 @@ from http import HTTPStatus
 from typing import Iterable
 
 from fastapi import APIRouter, HTTPException, Depends
+from pydantic import ValidationError
 from sqlmodel import Session
 from app.models.User import User, UserCreate, UserUpdate
 from app.database.engine import get_session
@@ -26,13 +27,20 @@ def get_users(session: Session = Depends(get_session)) -> Iterable[User]:
 
 @router.post("/", status_code=HTTPStatus.CREATED)
 def create_user(user: User, session: Session = Depends(get_session)) -> User:
-    UserCreate.model_validate(user.model_dump())
+    try:
+        UserCreate.model_validate(user.model_dump())
+    except ValidationError:
+        raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY)
     return users.create_user(session, user)
 
 @router.patch("/{user_id}", status_code=HTTPStatus.OK)
 def update_user(user_id: int, user: User, session: Session = Depends(get_session)) -> User:
     if user_id < 1:
         raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY, detail="Invalid user id")
+
+    if not users.get_user(session, user_id):
+        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
+
     UserUpdate.model_validate(user.model_dump())
     return users.update_user(session, user_id, user)
 
@@ -40,7 +48,13 @@ def update_user(user_id: int, user: User, session: Session = Depends(get_session
 def delete_user(user_id: int, session: Session = Depends(get_session)):
     if user_id < 1:
         raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY, detail="Invalid user id")
+
+    user = users.get_user(session, user_id)
+    if not user:
+        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
+
     users.delete_user(session, user_id)
     return {"message": "User deleted"}
+
